refactor : selectbox style 추출
