+++ src/components/common/selectBox/SelectBox.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useRef, useState } from "react";
 import { MOUSE_EVENT } from "../../../libs/interface/typeEvent";
-import { css, styled } from "styled-components";
 
 import ArrowIcon from "../../../assets/icon/icon_arrow_down.svg";
 import SelectItem from "./SelectItem";
 import { SelectType } from "../../../pages/myCoupon/UseCoupon";
+import { Arrow, Container, ItemList, SelectWrapper } from "./SelectBox.styled";
 
 const SelectBox = ({ styletype = "primary", selected, setSelected, data, placeholder, nulltext }: SelectType) => {
   const [isOpen, setIsOpen] = useState<boolean>(false);
@@ -55,76 +55,3 @@ const SelectBox = ({ styletype = "primary", selected, setSelected, data, placeho
 };
 
 export default SelectBox;
-
-type styleType = "primary" | "secondary";
-
-const Common = {
-  primary: css`
-    background-color: #f4f4f4;
-    font-size: 14px;
-  `,
-  secondary: css`
-    background-color: white;
-    box-shadow: 0 0 0 1px var(--gray200-color) inset;
-    font-size: 13px;
-  `,
-};
-
-const BoxType = {
-  primary: css`
-    margin-top: 15px;
-
-    padding: 13px 20px;
-  `,
-  secondary: css`
-    padding: 8.5px 10px;
-  `,
-};
-
-const ULType = {
-  primary: css`
-    top: 59px;
-    max-height: calc(40px * 5 + 0.5px * 5);
-  `,
-  secondary: css`
-    top: 50px;
-    max-height: calc(39px * 5 + 0.5px * 5);
-  `,
-};
-
-const SelectWrapper = styled.div<{ $styletype: styleType }>`
-  ${({ $styletype }) => Common[$styletype]}
-  ${({ $styletype }) => BoxType[$styletype]}
-
-  box-sizing: border-box;
-
-  border-radius: 5px;
-  font-family: var(--font--Regular);
-
-  position: relative;
-`;
-
-const Arrow = styled.img<{ $isopen: string }>`
-  transition: all 0.5s ease-out;
-  transform: ${({ $isopen }) => ($isopen === "true" ? `rotateX(180deg);` : undefined)};
-`;
-
-const Container = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const ItemList = styled.ul<{ $styletype: styleType }>`
-  box-sizing: border-box;
-  ${({ $styletype }) => Common[$styletype]}
-  ${({ $styletype }) => ULType[$styletype]}
-  padding: 0 12px;
-  width: 100%;
-  overflow-y: auto;
-  position: absolute;
-  left: 0;
-  border-radius: 5px;
-
-  z-index: 9999;
-`;
